requirements.txt
flask==3.0.0
line-bot-sdk==3.10.0
python-dotenv==1.0.0
apscheduler==3.10.4
